Reorganize test classes to match the new main organization
