+++ app/src/androidTest/java/ch/epfl/sdp/drone3d/ui/TempTestActivityTest.kt
@@ -1,10 +1,9 @@
-package ch.epfl.sdp.drone3d
+package ch.epfl.sdp.drone3d.ui
 
 import androidx.test.espresso.intent.Intents
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
-import ch.epfl.sdp.drone3d.ui.TempTestActivity
 import org.junit.*
 import org.junit.Assert.assertEquals
 import org.junit.runner.RunWith
