+++ app/src/androidTest/java/ch/epfl/sdp/drone3d/ui/drone/DroneConnectActivityTest.kt
@@ -1,4 +1,4 @@
-package ch.epfl.sdp.drone3d
+package ch.epfl.sdp.drone3d.ui.drone
 
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions
@@ -10,12 +10,11 @@ import androidx.test.espresso.matcher.ViewMatchers
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
+import ch.epfl.sdp.drone3d.R
 import ch.epfl.sdp.drone3d.drone.DroneInstanceProvider.getIP
 import ch.epfl.sdp.drone3d.drone.DroneInstanceProvider.getPort
 import ch.epfl.sdp.drone3d.drone.DroneInstanceProvider.isConnected
 import ch.epfl.sdp.drone3d.drone.DroneInstanceProvider.isSimulation
-import ch.epfl.sdp.drone3d.ui.drone.ConnectedDroneActivity
-import ch.epfl.sdp.drone3d.ui.drone.DroneConnectActivity
 import junit.framework.Assert.assertEquals
 import org.junit.*
 import org.junit.runner.RunWith
